+++ CMakeLists.txt
@@ -37,11 +37,13 @@ include(StaticAnalyzers)
 include(CPM)
 
 CPMAddPackage("gh:ambroise-leclerc/networking-ts-impl@1.12.0a")
+#CPMAddPackage("gh:ambroise-leclerc/CEF@118.7.1c")
+CPMAddPackage(
+  NAME cef
+  GITHUB_REPOSITORY ambroise-leclerc/CEF
+  GIT_TAG 17-cef-v137
+)
 
-# Only add CEF for non-Apple platforms for now due to architecture issues
-if(NOT APPLE)
-    CPMAddPackage("gh:ambroise-leclerc/CEF@118.7.1b")
-endif()
 
 # Create our own CEF wrapper library target
 set(CEF_ROOT "${CMAKE_BINARY_DIR}/_deps/cef_binaries-src")
@@ -85,6 +87,11 @@ endif()
 
 target_link_libraries(WebFront INTERFACE WebFront_options WebFront_warnings)
 target_include_directories(WebFront INTERFACE ${CMAKE_SOURCE_DIR}/include)
+# Ajouter les headers CEF à WebFront comme headers système pour éviter les warnings
+if(EXISTS "${CEF_ROOT}")
+  target_include_directories(WebFront SYSTEM INTERFACE "${CEF_ROOT}")
+  target_include_directories(WebFront SYSTEM INTERFACE "${CEF_ROOT}/include")
+endif()
 include_directories(SYSTEM ${networking-ts-impl_SOURCE_DIR}/include)
 
 option(BUILD_SHARED_LIBS "Enable compilation of shared libraries" OFF)
